compositor-drm: launch without input devices
During demonstrations of Automotive Grade Linux (AGL), GENIVI Development
Platform (GDP), Tizen and other platforms relying on Weston it is very
convenient to launch it even if there are no input devices such as a
keyboard or a mouse. This fix is particularly useful for Raspberry Pi
which is a frequently used and a very popular board in the community.

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>